Add FlowerShop and Products
Added 'display' attribute to FlowerShop with initial
instructional message to user on uploading an order file. To be
displayed on #run.

Added 'products' attribute to FlowerShop. This attribute holds
data on what products are on sale. i.e. flower types, batch sizes
and products. Products and .all added to allow mock to work in
test.
All 3 tests for FlowerShop.new passing.