Update #unfillable? algorithm
Previous algorithm does not work when the order is made up of the
5 and 3 only. For example, test failed when order was 13.
It should pass as 13 can be made up of 2 x 5 and 1 x 3.
The algorithm is similar but if the first run doesn't work it
drops off the largest in the bundle and runs through the modulo
operation again. In this case 9 was removed so that [5,3] remains.
This is not very efficient especially when multi-line orders are
added later, however, I think it's a fair assumption to make that
orders won't be in the hundres or thousands. So maybe it does
not matter.
All tests passing.