Fix bug in OrderValidator
The algorithm for identifying if an order is fillable did not
validate all cases. This bug was identified when a test was run
for 11 Tulips, which can be made up of 1 x 5 and 2 x 3 bundles

The new algorithm has the following steps:
- Find all possible combinations that are less than or equal to
the ordered amount. (#less_than_or_equal_to_combinations)
- Find combinations that match the ordered amount, (#find_matches)
- If matches exist then false is returned for #unfillable?

The algorithm was also extracted to a Fillable module