Finish implementation of Invoicer
All tests passing.
Removed total_value attribute because it's not needed. Tests
removed too.
Tests written to replicate output of specification document but
unable to replicate due to type in doc. Total value for 2 x 5
bundle of T58 was quoted as $9.95, which is the value of one.
Value of 2 should be $19.90

All tests passing.